sorted some more bits
There are times when only the expression d'uh will cover it. I wasn't
feeling the love for Bootstrap's fluid grid thingy as it didn't seem to
make any appreciable difference until I realised that the path to the
link in the head pointed to a folder that didn't exist. Oh well, it
plays much more nicely now.